+++ NEWS.md
@@ -1,4 +1,4 @@
-# power4mome 0.1.1.75
+# power4mome 0.1.1.77
 
 - Improve the function for extending the
   initial interval before doing a
@@ -447,6 +447,19 @@
 - Fixed the help for the argument
   `mode` of `q_power_mediation()`
   and friends. (0.1.1.75)
+
+- If the model is a multigroup model,
+  descriptive statistics for the
+  generated data are printed by
+  group, using `psych::describeBy()`.
+  (0.1.1.76)
+
+- Response proportions will be printed
+  with `data_long = TRUE` if a variable
+  has 10 or fewer unique values
+  (e.g., ordinal variables),
+  using `psych::responseFrequency()`.
+  (0.1.1.77)
 
 # power4mome 0.1.1